+++ samples/Java/samples/rfq/SampleRfq.java
@@ -46,8 +46,8 @@ public class SampleRfq extends SimpleWrapper {
    private double m_bidPrice = 0;
    private double m_askPrice = 0;
 
-   private int m_bidSize = 0;
-   private int m_askSize = 0;
+   private long m_bidSize = 0;
+   private long m_askSize = 0;
 
    private int m_receivedTicks = 0;
 
@@ -385,7 +385,7 @@ public class SampleRfq extends SimpleWrapper {
       }
    }    
 
-   public void tickSize(int tickerId, int field, int size) {
+   public void tickSize(int tickerId, int field, long size) {
        TickType tick = TickType.get(field);
        consoleMsg("tickSize:" + tickerId + " field:" + field +
             " (" + tick.field() + ") value:" + size);
